refactor main's theme and myApp
