+++ lib/main.dart
@@ -7,9 +7,12 @@ import 'package:family_altar/screens/reader/bloc/reading_bloc.dart';
 import 'package:family_altar/screens/reader/reader_screen.dart';
 import 'package:family_altar/screens/settings/settings_screen.dart';
 import 'package:family_altar/storage/local_reading_storage.dart';
-import 'package:family_altar/theme/bloc/theme_provider.dart';
+import 'package:family_altar/theme/bloc/theme_bloc.dart';
+import 'package:family_altar/theme/bloc/theme_event.dart';
+import 'package:family_altar/theme/bloc/theme_state.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:go_router/go_router.dart';
 
 const String appTitle = 'The Family Altar - Tim Dodd';
@@ -69,15 +72,54 @@ void main() async {
           providers: [
             // Global ReadingBloc
             BlocProvider(
-              create: (_) => ReadingBloc(readingRepository: readingRepository)
-                ..add(LoadReadingEvent(date: DateTime.now())),
+              create:
+                  (_) =>
+                      ReadingBloc(readingRepository: readingRepository)
+                        ..add(LoadReadingEvent(date: DateTime.now())),
+            ),
+            BlocProvider(
+              create: (_) => ThemeBloc()..add(const ThemeInitializeEvent()),
             ),
           ],
-          child: ThemeProvider(
-            router: _router,
-          ), 
+          child: MyApp(router: _router),
         ),
       ),
     ),
   );
+}
+
+class MyApp extends StatelessWidget {
+  const MyApp({required this.router, super.key});
+
+  final GoRouter router;
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocBuilder<ThemeBloc, ThemeState>(
+      builder: (context, themeState) {
+        return MaterialApp.router(
+          routerConfig: router,
+          title: 'The Family Altar',
+          debugShowCheckedModeBanner: false,
+          theme: themeState.lightTheme,
+          darkTheme: themeState.darkTheme,
+          themeMode: themeState.themeMode,
+          locale: TranslationProvider.of(context).flutterLocale,
+          supportedLocales: AppLocaleUtils.supportedLocales,
+          localizationsDelegates: const [
+            GlobalMaterialLocalizations.delegate,
+            GlobalWidgetsLocalizations.delegate,
+            GlobalCupertinoLocalizations.delegate,
+          ],
+          builder: (context, child) {
+            final mediaQuery = MediaQuery.of(context);
+            return MediaQuery(
+              data: mediaQuery.copyWith(textScaler: TextScaler.noScaling),
+              child: child!,
+            );
+          },
+        );
+      },
+    );
+  }
 }
