--- lib/theme/bloc/theme_provider.dart
@@ -1,53 +1,0 @@
-// theme/bloc/theme_provider.dart
-import 'package:family_altar/i18n/strings.g.dart';
-import 'package:family_altar/theme/bloc/theme_bloc.dart';
-import 'package:family_altar/theme/bloc/theme_event.dart';
-import 'package:family_altar/theme/bloc/theme_state.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:go_router/go_router.dart';
-
-class ThemeProvider extends StatelessWidget {
-  const ThemeProvider({
-    required this.router,
-    super.key,
-  });
-
-  final GoRouter router;
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocProvider(
-      create: (_) => ThemeBloc()..add(const ThemeInitializeEvent()),
-      child: BlocBuilder<ThemeBloc, ThemeState>(
-        builder: (context, themeState) {
-          return MaterialApp.router(
-            routerConfig: router,
-            title: 'The Family Altar',
-            debugShowCheckedModeBanner: false,
-            theme: themeState.lightTheme,
-            darkTheme: themeState.darkTheme,
-            themeMode: themeState.themeMode,
-            locale: TranslationProvider.of(context).flutterLocale,
-            supportedLocales: AppLocaleUtils.supportedLocales,
-            localizationsDelegates: const [
-              GlobalMaterialLocalizations.delegate,
-              GlobalWidgetsLocalizations.delegate,
-              GlobalCupertinoLocalizations.delegate,
-            ],
-            builder: (context, child) {
-              final mediaQuery = MediaQuery.of(context);
-              return MediaQuery(
-                data: mediaQuery.copyWith(
-                  textScaler: TextScaler.noScaling
-                ),
-                child: child!,
-              );
-            },
-          );
-        },
-      ),
-    );
-  }
-}
